Trigger playlists with NFC tags
The playlist is named after the NFC tag's ID, and is expected to be
in the `nfc/` directory of the current working directory (an error
showing the ID will appear in the player interface when presented with
a new NFC tag). It is probably a good idea to run `deck spin` from the
root of the music library, and symlink the filename to a playlist with
a more memorable filename.

Also adds previously forgotten 'skip' command to the CLI.